// admin/src/api/generated/schemas/adminAdminSwapProviderItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminAdminSwapProviderItem = {
  id?: string;
  provider_code?: string;
  provider_name?: string;
  description?: string;
  logo_url?: string;
  is_enabled?: boolean;
  min_swap_amount_usd?: string;
  max_swap_amount_usd?: string;
  default_slippage?: string;
  max_slippage?: string;
  fee_rate?: string;
  config_json?: string;
  created_at?: string;
  updated_at?: string;
};